+++ tests/test_baseflumine.py
@@ -232,10 +232,9 @@ class BaseFlumineTest(unittest.TestCase):
         mock_strategy.stream_ids = [1, 2, 3]
         self.base_flumine.strategies = [mock_strategy]
         mock_market = mock.Mock(closed=False, elapsed_seconds_closed=None)
-        mock_market.market_book.streaming_unique_id = 2
         self.base_flumine.markets._markets = {"1.23": mock_market}
         mock_event = mock.Mock()
-        mock_market_book = mock.Mock(market_id="1.23")
+        mock_market_book = mock.Mock(market_id="1.23", streaming_unique_id=2)
         mock_event.event = mock_market_book
         self.base_flumine._process_close_market(mock_event)
         mock_market.close_market.assert_called_with()